+++ mycodo/utils/system_pi.py
@@ -27,7 +27,7 @@ if logging.getLevelName(logging.getLogger().getEffectiveLevel()) == 'INFO':
     logger.setLevel(logging.INFO)
 
 
-def parse_custom_option_values(controllers):
+def parse_custom_option_values(controllers, dict_inputs):
     # Check if controllers is iterable or a single controller
     try:
         _ = iter(controllers)
@@ -47,6 +47,16 @@ def parse_custom_option_values(controllers):
                 else:
                     value = each_option.split(',')[1]
                 custom_options_values[each_controller.unique_id][option] = value
+
+        if 'custom_options' in dict_inputs[each_controller.device]:
+            dict_custom_options = dict_inputs[each_controller.device]['custom_options']
+        else:
+            dict_custom_options = {}
+        for each_option in dict_custom_options:
+            if ('id' in each_option and
+                    'default_value' in each_option and
+                    each_option['id'] not in custom_options_values[each_controller.unique_id]):
+                custom_options_values[each_controller.unique_id][each_option['id']] = each_option['default_value']
 
     return custom_options_values